fix(amount type)
all amount types of trade changed to int

+++ models/trade/closeTradeRequest.go
@@ -3,5 +3,5 @@ package trade
 // ClosedTradeRequest represents the request to close a trade
 type ClosedTradeRequest struct {
 	OpenTradeID int     `json:"id"`
-	Amount      float64 `json:"amount"`
+	Amount      int 	`json:"amount"`
 }
